pyproject.toml
[tool.poetry]
name = "autoresearch-llama3-crew"
version = "0.1.0"
description = "This is a crew for automating literature review process."
authors = ["Junhui Cho <jh00@bu.edu>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
python-dotenv = "^1.0.0"
opencv-python = "4.8.1.78"
google-search-results = "2.4.2"
crewai-tools = "0.4.0"
pypdf2 = "^3.0.0"
langchain-groq = "^0.0.1"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"